+++ ui/src/pages/CreateRecipePage.tsx
@@ -3,7 +3,7 @@ import { useDispatch } from "react-redux";
 import { useHistory } from "react-router-dom";
 import { IRecipeModel } from "../../../src/db/recipe";
 import { IRecipe } from "../../../src/types";
-import { useApi } from "../apiClient";
+import { useApi } from "../hooks/useApi";
 import RecipeForm from "../features/recipe/RecipeForm";
 import RecipeHeader from "../features/recipe/RecipeHeader";
 import { putRecipe } from "../features/recipe/RecipeSlice";
