+++ ui/src/pages/EditRecipePage.tsx
@@ -3,7 +3,7 @@ import { useDispatch, useSelector } from "react-redux";
 import { useHistory, useParams } from "react-router-dom";
 import { IRecipeModel } from "../../../src/db/recipe";
 import { IRecipe } from "../../../src/types";
-import { useApi } from "../apiClient";
+import { useApi } from "../hooks/useApi";
 import RecipeForm from "../features/recipe/RecipeForm";
 import RecipeHeader from "../features/recipe/RecipeHeader";
 import { putRecipe, selectRecipe } from "../features/recipe/RecipeSlice";
